Refactor booking functionality and update variable names

+++ website/booking.py
@@ -1,20 +1,20 @@
 from flask import Blueprint, redirect, render_template, request, url_for
 from flask_login import login_required, current_user
-from sqlalchemy import delete
-from .models import Booking
+from .models import Booking, SERVICES
 from . import db
 
 booking = Blueprint('booking', __name__)
 
-@booking.route('/book-ticket', methods=['GET', 'POST'])
+@booking.route('/book-appointment', methods=['GET', 'POST'])
 @login_required
-def book_ticket():
+def book_appointment():
     if request.method == "POST":
-        date = request.form.get('date')
-        time = request.form.get('time')
+        date = request.form.get('appointmentDate')
+        time = request.form.get('appointmentTime')
         service = request.form.get('service')
+        price = SERVICES[service]
 
-        new_booking = Booking(date=date, time=time, service=service, user_id=current_user.id)
+        new_booking = Booking(date=date, time=time, service=service, price=price, user_id=current_user.id)
         db.session.add(new_booking)
         db.session.commit()
 
@@ -25,12 +25,12 @@ def book_ticket():
 @login_required
 def my_bookings():
     bookings = Booking.query.filter_by(user_id=current_user.id).all()
+
+    if request.method == "POST":
+        booking_id = request.form.get('cancel-booking')
+        booking = Booking.query.get(int(booking_id))
+        db.session.delete(booking)
+        db.session.commit()
+        return redirect(url_for('booking.my_bookings'))
     return render_template("my_bookings.html", user=current_user, bookings=bookings)
 
-@booking.route('/delete-booking/<int:id>')
-@login_required
-def delete_booking(id):
-    booking = Booking.query.filter_by(id=id).first()
-    db.session.delete(booking)
-    db.session.commit()
-    return redirect(url_for('booking.my_bookings'))
